refactor: export downloadJsonAsFile as standalone helper
