+++ packages/core-app-elements/src/main.ts
@@ -68,4 +68,5 @@ export { default as ListItemTask } from '#ui/lists/ListItemTask'
 export { default as TableData } from '#ui/tables/TableData'
 
 // Helpers
+export { downloadJsonAsFile } from '#helpers/downloadJsonAsFile'
 export { formatDate } from '#helpers/date'
